modularize visualize codes
