+++ src/cache/visualize.py
@@ -1,10 +1,12 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
+import json
 import math
 import os
 import pickle
 import re
+from collections import defaultdict
 from glob import glob
 from pprint import pformat
 from typing import Dict, List
@@ -193,6 +195,91 @@ def read_tuning_data(prefix):
         M, N = MN
         sorted_fig_data[(M, N)] = sort_by_key(fig_data[MN], "K")
     return sorted_fig_data
+
+
+def plot_single_pe_vs_k(m, n, k_data, kernel_dir):
+    """
+    Create a single plot of PE utilization vs K for a specific (M, N) pair.
+
+    Parameters:
+    -----------
+    m, n : int
+        M and N dimensions.
+    k_data : dict
+        Dictionary mapping K values to their respective PE utilization.
+    kernel_dir : str
+        Directory to save the plot.
+    """
+    # Sort the K values
+    k_values = sorted(k_data.keys())
+    pe_util_values = [k_data[k] for k in k_values]
+
+    plt.figure(figsize=(10, 6))
+    plt.plot(k_values, pe_util_values, "o-", linewidth=2)
+    plt.title(f"PE Utilization vs. K for M={m}, N={n}")
+    plt.xlabel("K")
+    plt.ylabel("PE Utilization (%)")
+    plt.grid(True)
+
+    # Set x-ticks to powers of 2 within the range of k_values
+    min_k, max_k = min(k_values), max(k_values)
+
+    # Generate powers of 2 that lie within our range
+    powers = []
+    power = 1
+    while power <= max_k:
+        if power >= min_k:
+            powers.append(power)
+        power *= 2
+
+    # If there are power of 2 values in our range, use them as ticks
+    if powers:
+        plt.xticks(powers, [str(x) for x in powers])
+
+    # Format y-axis as percentages
+    plt.gca().yaxis.set_major_formatter(plt.FuncFormatter(lambda x, _: f"{x*100:.0f}%"))
+
+    plt.tight_layout()
+    plt.savefig(f"{kernel_dir}/pe_utilization_M{m}_N{n}.png", dpi=400)
+    plt.close()
+
+
+def plot_pe_vs_k(kernel):
+    """
+    Create plots of PE utilization vs K for each unique (M, N) pair.
+
+    Parameters:
+    -----------
+    kernel : object
+        The kernel object containing func_name attribute.
+    """
+    # Dictionary to hold PE utilization data for each (M, N, K) combination
+    pe_utilization_data = defaultdict(dict)
+
+    # Check if the kernel directory exists
+    kernel_dir = f"{TUNED_NKI_CACHE_DIR}/{kernel.func_name}"
+    assert os.path.exists(kernel_dir), f"Directory not found: {kernel_dir}"
+
+    # Scan the directories to find all MNK combinations
+    for dirname in os.listdir(kernel_dir):
+        # Extract M, N, K values from directory name
+        match = re.match(r"M(\d+)-N(\d+)-K(\d+)", dirname)
+        if match:
+            m, n, k = map(int, match.groups())
+
+            # Read the perf_metrics.json file
+            json_path = os.path.join(kernel_dir, dirname, "perf_metrics.json")
+            if os.path.exists(json_path):
+                with open(json_path, "r") as f:
+                    data = json.load(f)
+                    # Find the configuration with the lowest latency
+                    best_config = min(data["results"], key=lambda x: x["latency"])
+                    # Store the PE utilization of the best configuration
+                    pe_utilization_data[(m, n)][k] = best_config["pe_utilization"]
+
+    # Create plots for each (M, N) pair
+    for (m, n), k_data in pe_utilization_data.items():
+        plot_single_pe_vs_k(m, n, k_data, kernel_dir)
 
 
 if __name__ == "__main__":
